+++ sbgECom/src/protocol/sbgEComProtocol.h
@@ -51,7 +51,7 @@ typedef struct _SbgEComProtocol
 {
 	SbgInterface		*pLinkedInterface;							/*!< Associated interface used by the protocol to read/write bytes. */
 	uint8				 rxBuffer[SBG_ECOM_MAX_BUFFER_SIZE];		/*!< The reception buffer. */
-	size_t				 rxBufferSize;								/*!< The current reception buffer size in bytes. */
+	uint32				 rxBufferSize;								/*!< The current reception buffer size in bytes. */
 } SbgEComProtocol;
 
 //----------------------------------------------------------------------//
@@ -82,7 +82,7 @@ SbgErrorCode sbgEComProtocolClose(SbgEComProtocol *pHandle);
  * \param[in]	size					Size in bytes of the data payload (less than 4086).
  * \return								SBG_NO_ERROR if the frame has been sent.
  */
-SbgErrorCode sbgEComProtocolSend(SbgEComProtocol *pHandle, uint8 msgClass, uint8 msg, const void *pData, size_t size);
+SbgErrorCode sbgEComProtocolSend(SbgEComProtocol *pHandle, uint8 msgClass, uint8 msg, const void *pData, uint32 size);
 
 /*!
  * Try to receive a frame from the device and returns the cmd, data and size of data field.
@@ -98,7 +98,7 @@ SbgErrorCode sbgEComProtocolSend(SbgEComProtocol *pHandle, uint8 msgClass, uint8
  *										SBG_NULL_POINTER if an input parameter is NULL.<br>
  *										SBG_BUFFER_OVERFLOW if the received frame payload couldn't fit into the pData buffer.
  */
-SbgErrorCode sbgEComProtocolReceive(SbgEComProtocol *pHandle, uint8 *pMsgClass, uint8 *pMsg, void *pData, size_t *pSize, size_t maxSize);
+SbgErrorCode sbgEComProtocolReceive(SbgEComProtocol *pHandle, uint8 *pMsgClass, uint8 *pMsg, void *pData, uint32 *pSize, uint32 maxSize);
 
 //----------------------------------------------------------------------//
 //- Frame generation to stream buffer                                  -//
